// src/typesOrInterface/types.ts
export type listType = {
	name: string;
	arrowAlt: string;
	alt: string;
};

export type picturesType = {
	cube?: {
		src: string;
		width: number;
		height: number;
		blurHeight: number;
		blurWidth: number;
	};
	discont?: {
		src: string;
		width: number;
		height: number;
		blurHeight: number;
		blurWidth: number;
	};
	icon?: {
		src: string;
		width: number;
		height: number;
		blurHeight: number;
		blurWidth: number;
	};
	key?: {
		src: string;
		width: number;
		height: number;
		blurHeight: number;
		blurWidth: number;
	};
	wallet?: {
		src: string;
		width: number;
		height: number;
		blurHeight: number;
		blurWidth: number;
	};
	question?: {
		src: string;
		width: number;
		height: number;
		blurHeight: number;
		blurWidth: number;
	};
};